requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
cryptography==45.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-redis==5.4.0
django-widget-tweaks==1.5.0
fonttools==4.58.4
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.4.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.28.0
pyhanko-certvalidator==0.27.0
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.2
redis==6.1.0
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1